Changed addr type from ulong to uint
This is done due to new revelations in the ptrace man page. PT_READ_D/I reads exactly "int" of data, so the long was kind of overestimated.

+++ src/runtime/breakpoint/breakpoint.h
@@ -7,7 +7,7 @@ struct breakpoint_t
 	char *path; /* path to file in which is the breakpoint */
 	long orig; /* original instruction, before substitution for 0xCC */
 	long oxcc; /* original instruction, with first byte modified to 0xCC */
-	unsigned long addr; /* address of the address in executable corresponding to 
+	unsigned int addr; /* address of the address in executable corresponding to 
 	                     * line number */
 
 	struct breakpoint_t *next;  /* pointer to next node in linked list */
